+++ gsrest/service/stats_service.py
@@ -2,7 +2,7 @@ from openapi_server.models.currency_stats import CurrencyStats
 from gsrest.db.util import tagstores
 
 
-async def get_currency_statistics(request, currency, version=None):
+async def get_currency_statistics(request, currency):
     db = request.app['db']
     result = await db.get_currency_statistics(currency)
     if result is None:
@@ -30,3 +30,12 @@ async def get_currency_statistics(request, currency, version=None):
             no_tagged_addresses=int(no_tagged_addresses),
             timestamp=result['timestamp']
         )
+
+
+async def get_no_blocks(request, currency):
+    db = request.app['db']
+    result = await db.get_currency_statistics(currency)
+    if result is None:
+        raise SystemError('statistics for currency {} not found'
+                          .format(currency))
+    return result['no_blocks']